[refactor] Move Composables in their respective screen directories
* Rename FixtureList to FixturesListView for naming consistency throughout the project (similar to playersListView).
