+++ app/src/main/java/dev/johnoreilly/fantasypremierleague/di/AppModule.kt
@@ -1,6 +1,6 @@
 package dev.johnoreilly.fantasypremierleague.di
 
-import dev.johnoreilly.fantasypremierleague.ui.FantasyPremierLeagueViewModel
+import dev.johnoreilly.fantasypremierleague.ui.players.FantasyPremierLeagueViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
