+++ app/src/main/java/dev/johnoreilly/fantasypremierleague/ui/global/Theme.kt
@@ -1,4 +1,4 @@
-package dev.johnoreilly.fantasypremierleague.ui
+package dev.johnoreilly.fantasypremierleague.ui.global
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
@@ -21,11 +21,8 @@ private val LightColorPalette = lightColors(
 
 @Composable
 fun FantasyPremierLeagueTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable() () -> Unit) {
-    val colors = if (darkTheme) {
-        DarkColorPalette
-    } else {
-        LightColorPalette
-    }
+    val colors = if (darkTheme) DarkColorPalette
+    else LightColorPalette
 
     MaterialTheme(
         colors = colors,
