+++ app/src/main/java/dev/johnoreilly/fantasypremierleague/ui/players/FantasyPremierLeagueViewModel.kt
@@ -1,4 +1,4 @@
-package dev.johnoreilly.fantasypremierleague.ui
+package dev.johnoreilly.fantasypremierleague.ui.players
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.*
